Actualizare hardware (#369)
* GreenGuardian

The GreenGuardian project is a smart plant monitoring system designed to effortlessly track and maintain optimal conditions for plants. The system incorporates various environmental sensors to monitor parameters such as soil moisture, ambient temperature, light intensity, humidity, and detect vibrations caused by bugs or disturbances. The motivation behind the project stems from experiences of neglecting plant care and the need for a more convenient method for plant maintenance

* Update index.md

* Hardware

I have added the hardware part and in the log I said what is missing in order to implement the project.